Доработки в проекте
