+++ java/lang/annotations.xml
@@ -1,0 +1,5 @@
+<root>
+    <item name='java.lang.ClassCastException'>
+        <annotation name='java.lang.Deprecated'/>
+    </item>
+</root>
