+++ pom.xml
@@ -32,8 +32,14 @@
         <project.java.version>1.8</project.java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <project.reporting.outputEncoding>UTF-8</project.reporting.outputEncoding>
-
     </properties>
 
-
+    <dependencies>
+       <dependency>
+           <groupId>org.projectlombok</groupId>
+           <artifactId>lombok</artifactId>
+           <version>1.18.12</version>
+           <scope>provided</scope>
+       </dependency>
+    </dependencies>
 </project>
