+++ src/main/java/ru/exapmple/Limit.java
@@ -1,7 +1,11 @@
-package interfaces;
+package ru.exapmple;
+
+import lombok.SneakyThrows;
 
 public class Limit {
-        public static void main(String [] args) {
+
+        @SneakyThrows
+                public static void main(String [] args) {
                 Checking c = new Checking(101);
                 System.out.println("Остаток 300 рублей...");
                 c.limit(300.00);
@@ -11,9 +15,12 @@ public class Limit {
                         c.withdraw(100.00);
                         System.out.println("\nПеревод 400 рублей...");
                         c.withdraw(400.00);
-                }catch(InsufficientFundsException e) {
+                } catch (InsufficientFundsException e) {
                         System.out.println("Извините, но у Вас всего рублей " + e.getAmount());
                         e.printStackTrace();
+
                 }
+
+
         }
 }
