+++ src/main/java/ru/exapmple/MyException.java
@@ -1,4 +1,4 @@
-package interfaces;
+package ru.exapmple;
 
 public class MyException extends Exception{
