+++ src/main/java/ru/exapmple/Operation.java
@@ -1,4 +1,4 @@
-package interfaces;
+package ru.exapmple;
 
 public interface Operation {
     void payOperation();
