+++ src/main/java/ru/exapmple/User.java
@@ -1,0 +1,9 @@
+package ru.exapmple;
+
+public class User  {
+    private int id;
+    public User(int id) {
+        this.id = id;
+    }
+
+}
