+++ src/test/java/app/Application.java
@@ -1,0 +1,30 @@
+package app;
+
+import lombok.EqualsAndHashCode;
+import ru.exapmple.Operation;
+
+import java.awt.*;
+@EqualsAndHashCode
+
+public class Application implements Operation {
+     public void pay () {
+         System.out.println("Payment numberPhone");
+         //Использовать ArrayList, HashMap, Map или EqualsAndHashCode)
+     }
+        private int serverPort = 1234;
+
+        private int Hashtable = 5;
+
+        private String ipAddr = "74.125.225.196";
+
+        private static String numberPhone;
+
+     @Override
+     public void payOperation() {
+
+     }
+
+
+ }
+
+
