+++ src/test/java/app/Server.java
@@ -1,13 +1,30 @@
-package interfaces;
+package app;
 
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.Setter;
 
+import java.util.HashMap;
+import java.util.Map;
+
+@Getter
+@Setter
+@EqualsAndHashCode
 
 public class Server {
+    private Map<User, Integer> items = new HashMap<>();
+    // какая-то логика
+    //for......
+    //System out print
+
+
 
     public static void main(String[] args) {
 // сервер реализует с приложением и клиентом (слушает запросы, распознает переданные параметры и значения,
  //       корректно отвечает на них
     }
 
-    }
+
+}
+
