--- src/test/java/interfaces/Application.java
@@ -1,23 +1,0 @@
-package interfaces;
-
- public class Application implements Operation {
-     public void pay () {
-
-         System.out.println("Payment numberPhone");
-     }
-        private int serverPort = 1234;
-
-        private int Hashtable = 5;
-
-        private String ipAddr = "74.125.225.196";
-
-
-
-     public void payOperation() {
-
-     }
-
-
- }
-
-
